+++ src/OginiServiceProvider.php
@@ -22,7 +22,7 @@ class OginiServiceProvider extends ServiceProvider
     /**
      * Package version.
      */
-    const VERSION = '1.0.5';
+    const VERSION = '1.0.6';
     /**
      * Register any application services.
      *